+++ engines/plumbers/configure.engine
@@ -1,3 +1,3 @@
 # This file is included from the main "configure" script
 # add_engine [name] [desc] [build-by-default] [subengines] [base games] [deps]
-add_engine mario "Luigi" no "" "" ""
+add_engine plumbers "Plumbers Don't Wear Ties" no "" "" ""
